Made the MCP client 1. Connect to openai endpoint and make queries 2. Start and connect to MCP servers and know about all available tools 3. Handle LLM response where the LLM wants us to make a tool call.
The MCP client acts as the broker between the LLM endpoint and the MCP servers. It is also responsible for maintaining the proper message list (with tool call requests and responses) we send to the LLM endpoint so that the LLM always has context from previous conversation.

Also, changed conversations variable to messages since that seems like the standard when reading other code on github.